made path support specific for windows
now the script should check if system is windows or not.
If windows then it uses firefox binary path else dosen't use it.
This should make the script compatible with docker/linux/mac and easy for windows users.
Also increased the time.sleep to 8 seconds since I experienced some blocks with 5 second timeout.